// src/api/models/AutenticacionLoginInternoResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AutenticacionLoginInternoResponse } from './AutenticacionLoginInternoResponse';
export type AutenticacionLoginInternoResponseApiResponse = {
    success?: boolean;
    data?: AutenticacionLoginInternoResponse;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};