// src/api/models/RegistroClienteEmpresaResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { RegistroClienteEmpresaResponse } from './RegistroClienteEmpresaResponse';
export type RegistroClienteEmpresaResponseApiResponse = {
    success?: boolean;
    data?: RegistroClienteEmpresaResponse;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};