// src/api/models/RegistroClienteNaturalResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { RegistroClienteNaturalResponse } from './RegistroClienteNaturalResponse';
export type RegistroClienteNaturalResponseApiResponse = {
    success?: boolean;
    data?: RegistroClienteNaturalResponse;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};